+++ src/main/java/eu/siacs/conversations/entities/Conversation.java
@@ -173,7 +173,7 @@ public class Conversation extends AbstractEntity implements Blockable, Comparabl
     public String findMostRecentRemoteDisplayableId() {
         final boolean multi = mode == Conversation.MODE_MULTI;
         synchronized (this.messages) {
-            for(final Message message : Lists.reverse(this.messages)) {
+            for (final Message message : Lists.reverse(this.messages)) {
                 if (message.getStatus() == Message.STATUS_RECEIVED) {
                     final String serverMsgId = message.getServerMsgId();
                     if (serverMsgId != null && multi) {
@@ -185,6 +185,21 @@ public class Conversation extends AbstractEntity implements Blockable, Comparabl
         }
         return null;
     }
+
+    public Message getLastEditableMessage() {
+        synchronized (this.messages) {
+            for (final Message message : Lists.reverse(this.messages)) {
+                if (message.isEditable()) {
+                    if (message.isGeoUri() || message.getType() != Message.TYPE_TEXT) {
+                        return null;
+                    }
+                    return message;
+                }
+            }
+        }
+        return null;
+    }
+
 
     public Message findUnsentMessageWithUuid(String uuid) {
         synchronized (this.messages) {
@@ -499,7 +514,7 @@ public class Conversation extends AbstractEntity implements Blockable, Comparabl
     @Override
     public int compareTo(@NonNull Conversation another) {
         return ComparisonChain.start()
-                .compareFalseFirst(another.getBooleanAttribute(ATTRIBUTE_PINNED_ON_TOP, false), getBooleanAttribute(ATTRIBUTE_PINNED_ON_TOP,false))
+                .compareFalseFirst(another.getBooleanAttribute(ATTRIBUTE_PINNED_ON_TOP, false), getBooleanAttribute(ATTRIBUTE_PINNED_ON_TOP, false))
                 .compare(another.getSortableTime(), getSortableTime())
                 .result();
     }