+++ src/main/java/eu/siacs/conversations/xmpp/jingle/WebRTCWrapper.java
@@ -370,12 +370,19 @@ public class WebRTCWrapper {
         }
     }
 
-    void setMicrophoneEnabled(final boolean enabled) {
+    boolean setMicrophoneEnabled(final boolean enabled) {
         final AudioTrack audioTrack = this.localAudioTrack;
         if (audioTrack == null) {
             throw new IllegalStateException("Local audio track does not exist (yet)");
         }
-        audioTrack.setEnabled(enabled);
+        try {
+            audioTrack.setEnabled(enabled);
+            return true;
+        } catch (final IllegalStateException e) {
+            Log.d(Config.LOGTAG, "unable to toggle microphone", e);
+            //ignoring race condition in case MediaStreamTrack has been disposed
+            return false;
+        }
     }
 
     boolean isVideoEnabled() {
@@ -502,7 +509,7 @@ public class WebRTCWrapper {
         final CameraEnumerator enumerator = getCameraEnumerator();
         final Set<String> deviceNames = ImmutableSet.copyOf(enumerator.getDeviceNames());
         for (final String deviceName : deviceNames) {
-            if (enumerator.isFrontFacing(deviceName)) {
+            if (isFrontFacing(enumerator, deviceName)) {
                 final CapturerChoice capturerChoice = of(enumerator, deviceName, deviceNames);
                 if (capturerChoice == null) {
                     return Optional.absent();
@@ -515,6 +522,14 @@ public class WebRTCWrapper {
             return Optional.absent();
         } else {
             return Optional.fromNullable(of(enumerator, Iterables.get(deviceNames, 0), deviceNames));
+        }
+    }
+
+    private static boolean isFrontFacing(final CameraEnumerator cameraEnumerator, final String deviceName) {
+        try {
+            return cameraEnumerator.isFrontFacing(deviceName);
+        } catch (final NullPointerException e) {
+            return false;
         }
     }